+++ src/datastore/index.ts
@@ -4,4 +4,5 @@ export * from '../datastore/ModelsDatastoreDB';
 export * from '../datastore/ModelsData';
 export * from '../datastore/MongoDB';
 export * from '../datastore/InstanceLocker';
+export * from '../datastore/QueryTranslator';